#6 Спринт 12 - Добавить функциональность для эндпоинтов mpa и genres: -Добавлены Get методы контроллер FilmController -Добавлены правки в ранее созданные DTO классы -Добавлены методы для получения данных из БД

+++ src/main/java/ru/yandex/practicum/filmorate/storage/mapper/RatingRowMapper.java
@@ -12,7 +12,7 @@ public class RatingRowMapper implements RowMapper<RatingDTO> {
     @Override
     public RatingDTO mapRow(ResultSet resultSet, int rowNum) throws SQLException {
         RatingDTO rating = new RatingDTO();
-        rating.setId(resultSet.getLong("id"));
+        rating.setId(resultSet.getInt("id"));
         rating.setRating(resultSet.getString("rating"));
         return rating;
     }
